Add files via upload
This is a xml decoder used for the PentesterLab exercise for the serialize badge. 
XMLDecoder is a Java class that creates object based on a XML message. If a malicious user can get an application to use arbitrary data in a call to the method readObject, she will instantly gain code execution on the server.